+++ src/main/java/racing/output/RacingOutput.java
@@ -1,4 +1,4 @@
-package racing.view.output;
+package racing.output;
 
 public class RacingOutput {
     public void print(String output) {
